feat: extract shared tree traversal utilities from optimizer passes
